requirements.txt
openai
botbuilder-integration-aiohttp>=4.14.0
pillow
azure-ai-projects
azure-identity
python-dotenv
jsonref
opencensus-ext-azure
autogen-agentchat
autogen-ext[openai,azure]
langgraph-0.2.74
langsmith
langchain-openai
langchain_community
langchain_experimental
ipython
websockets
botbuilder-core 
botbuilder-dialogs 
botbuilder-schema